+++ src/skydb_v2.test.js
@@ -12,7 +12,8 @@ const {
   MAX_ENTRY_LENGTH,
   JsonData,
 } = require("skynet-js");
-const { checkCachedDataLink, toHexString, decodeSkylink, getSettledValues } = require("./defaults");
+const { decodeSkylink, getSettledValues } = require("./defaults");
+const { checkCachedDataLink, toHexString } = require("./utils");
 
 // Generated with genKeyPairFromSeed("insecure test seed")
 const [publicKey, privateKey] = [